Refactor: DrawingCanvas 로직분리
